deprecate `AscentOwnedRow`
